Crate wasmer_toml

The wasmer.toml file format.

You’ll typically start by deserializing into a Manifest and inspecting its properties.

Re-exports

pub extern crate serde_cbor;

pub extern crate toml;

Modules

rust

Rust-specific annotations used to interoperate with external tools.

Structs

CommandV1Deprecated

Describes a command for a wasmer module.

CommandV2

An executable command.

FileCommandAnnotations

Annotations on disk.

Manifest

The manifest represents the file used to describe a Wasm package.

ManifestBuilder

Builder for Manifest.

Module

A file which may be executed by a Command. Sometimes also referred to as an “atom”.

Package

Metadata about the package.

PackageBuilder

Builder for Package.

WaiBindings

WitBindings

Enums

Abi

The ABI is a hint to WebAssembly runtimes about what additional imports to insert and how a module may be run.

Bindings

The interface exposed by a Module.

Command

CommandAnnotations

Annotations for a command.

FileKind

The different formats that FileCommandAnnotations can be saved in.

ImportsError

Errors that may occur when resolving Bindings imports.

ManifestBuilderError

Error type for ManifestBuilder

ManifestError

Errors that may occur while working with a Manifest.

ModuleReference

A reference to a module which may or may not come from another package.

PackageBuilderError

Error type for PackageBuilder

ValidationError

Errors that may be returned by Manifest::validate().

Statics

MANIFEST_FILE_NAME

The default name for the manifest file.